# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAphoto
  description: |-
    OpenAPI document for the WASAPhoto application
  version: 1.1.0
...
paths:
...
  /profiles/{userID}/followings/{targetUserID}:
    parameters:
    - $ref: "#/components/parameters/userID"
    - $ref: "#/components/parameters/targetUserID"
    put:
      summary: Follow a user
      description: Add the user that you want to follow into list of following
      tags: [ "User" ]
      operationId: followUser
      responses:
        "200":
          description: The user followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: Unfollow a user
      tags: [ "User" ]
      operationId: unfollowUser
      description: Unfollow the user with the given userID and delete it from the following
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /profiles/{userID}/bans/{targetUserID}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/targetUserID"
    put:
      summary: Ban a user
      tags: [ "User" ]
      operationId: banUser
      description: Ban the user with the given userID and delete it from the following
      responses:
        "200":
          description: User banned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Unban a user
      tags: [ "User" ]
      operationId: unbanUser
      description: Unban the user with the given userID and delete it from the following
      responses:
        "200":
          description: User unbanned successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /profiles/{userID}/posts/{photoID}/likes/{likeID}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/photoID"
      - $ref: "#/components/parameters/likeID"
    put:
      summary: Like a post
      tags: [ "Posts" ]
      operationId: likePhoto
      description: Like the post with the given photoID.
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Unlike a post
      tags: [ "Posts" ]
      operationId: unlikePhoto
      description: Unlike the post with the given photoID.
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /profiles/{userID}/posts/{photoID}/comments/{commentID}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/photoID"
      - $ref: "#/components/parameters/commentID"
    delete:
      summary: Delete a comment
      tags: [ "Posts" ]
      operationId: uncommentPhoto
      description: Delete the comment with the given commentID.
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    Username:
      title: Username
      description: The username of the user
      type: object
      properties:
        username:
          type: string
          pattern: "^[a-z0-9]{3,13}$"
          description: User's username
          example: "john123"
          minLength: 3
          maxLength: 13
    ImageData:
      title: Image
      description: The binary of the image (1-20Mb)
      type: string
      format: binary
      minLength: 1
      maxLength: 20971520
...
    UserID:
      type: integer
      description: The user id of the user.
      example: 100
    User:
      title: User
      description: The user object
      type: object
      properties:
        userID:
          $ref: "#/components/schemas/UserID"
        username:
          $ref: "#/components/schemas/Username"
        profilePicture:
          $ref: "#/components/schemas/ImageData"
...
    PhotoID:
      description: The ID of the post
      type: integer
      readOnly: true
      example: 100
...
    Message:
      title: Message
      description: This object represent a generic description of a message
      type: object
      readOnly: true
      properties:
        error:
          description: The message of the error
          type: string
          pattern: '\S.+'
          example: Username is already taken
          minLength: 0
          maxLength: 200
  parameters:
    userID:
      name: userID
      in: path
      required: true
      description: The user id of the user.
      schema:
        $ref: "#/components/schemas/UserID"
    photoID:
      name: photoID
      in: path
      description: The numeric photo id
      required: true
      schema:
        $ref: "#/components/schemas/PhotoID"
    commentID:
      name: commentID
      in: path
      description: The numeric comment id
      required: true
      schema:
        $ref: "#/components/schemas/Comment/properties/commentID"
    likeID:
      name: likeID
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UserID"
    targetUserID:
      name: targetUserID
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UserID"
  responses:
...
    Success:
      description: Success
      content:
        application/json:
          schema:
            description: Success message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 201 response
    Created:
      description: Created
      content:
        application/json:
          schema:
            description: Created message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 204 response
    NoContent:
      description: No content
      content:
        application/json:
          schema:
            description: No content message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 400 response
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            description: Bad request message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 401 response
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            description: Unauthorized message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 403 response
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            description: Forbidden message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
    # 404 response
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            description: Internal server error message
            type: object
            properties:
              message:
                $ref: "#/components/schemas/Message"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

# Apply the security global, to all the operations
security:
- bearerAuth: []